+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/shape/ShapeExporterBase.java
@@ -64,10 +64,15 @@ public abstract class ShapeExporterBase implements IShapeExporter {
     private final ColorTransform colorTransform;
 
     private boolean canUseSmoothing = true;
+    
+    protected int windingRule;
+    
+    
 
-    public ShapeExporterBase(int shapeNum, SWF swf, SHAPE shape, ColorTransform colorTransform) {
+    public ShapeExporterBase(int windingRule, int shapeNum, SWF swf, SHAPE shape, ColorTransform colorTransform) {
         this.shape = shape;
         this.colorTransform = colorTransform;
+        this.windingRule = windingRule;
 
         Cache<SHAPE, ShapeExportData> cache = swf.getShapeExportDataCache();
         ShapeExportData cachedData = cache.get(shape);