+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/DefineShape2Tag.java
@@ -83,4 +83,9 @@ public class DefineShape2Tag extends ShapeTag {
     public int getShapeNum() {
         return 2;
     }
+
+    @Override
+    public int getWindingRule() {
+        return WIND_EVEN_ODD;
+    }    
 }